requirements.txt
fastapi 
bson
pymongo
uvicorn
pyjwt
jose
python-jose-cryptodome
passlib[bcrypt]
pydantic[email]
python-multipart
bcrypt==4.0.1